fix: rename card-content features block
