+++ libs/blocks/src/lib/features/content-left/content-left.stories.tsx
@@ -4,7 +4,7 @@ import { Text } from '@deriv/quill-design';
 import ContentLeft from '.'
 
 const meta = {
-  title: 'Blocks/Features/Content-Left',
+  title: 'Blocks/Features/ContentLeft',
   component: ContentLeft,
   tags: ['autodocs'],
 } satisfies Meta<typeof ContentLeft>
